requirements.txt
loguru==0.7.2
openpyxl==3.1.5
playwright==1.48.0
python-dotenv==1.0.1
requests==2.32.3
web3==7.5.0
pyperclip==1.9.0
